Update index-suffix.html
there are three problems mentioned in #43 and but when i went to look into it, i found only a fourth, and not the three mentioned.

The problem i found is that http://didnotread.github.com/didnotread.org/index.html#twitter does not open the popup. this change should fix that.